requirements.txt
pyserial
numpy
platformio
prettytable